Use beam_lib to process chunks
Use the beam_lib module to process chunks, this takes care of size
calculations, padding and more for us.

Since we now use beam_lib to process chunks, drop the documentation of
the beam file format, as we no longer use this info. Replace it with
a link to the beam file format.